WIP. Clippy lints and project tweaks.
WIP: Anything else?

Fix a bunch of clippy lints (both new and those that somehow got missed
from earlier releases).

Change the optimisation level of Cargo's dev profile to 0, but leave
test at 2. This hopefully means that tests aren't unbearably slow, but
things like check and clippy are also faster.

+++ src/cli/srclist/convert.rs
@@ -182,7 +182,7 @@ fn convert<P: AsRef<Path>, S: AsRef<str>>(
                 "Using coarse channel frequencies [MHz]: {}",
                 coarse_chan_freqs
                     .iter()
-                    .map(|cc_freq_hz| format!("{:.2}", *cc_freq_hz as f64 / 1e6))
+                    .map(|cc_freq_hz| format!("{:.2}", *cc_freq_hz / 1e6))
                     .join(", ")
             );
